+++ lib/src/io/interface.dart
@@ -17,14 +17,22 @@ class Stderr {
   void flush() {}
 }
 
+/// An error thrown by [readFileAsBytes] and [readFileAsString].
+class FileSystemException {
+  String get message => null;
+}
+
 /// The standard error for the current process.
 Stderr get stderr => null;
 
 /// Returns whether or not stdout is connected to an interactive terminal.
 bool get hasTerminal => false;
 
-/// Reads the file at [path].
-String readFile(String path) => null;
+/// Reads the file at [path] as a list of bytes.
+List<int> readFileAsBytes(String path) => null;
+
+/// Reads the file at [path] as a UTF-8 encoded string.
+String readFileAsString(String path) => null;
 
 /// Returns whether a file at [path] exists.
 bool fileExists(String path) => null;