+++ lib/src/io/node.dart
@@ -2,11 +2,13 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'dart:typed_data';
+
 import 'package:js/js.dart';
 
 @JS()
 class _FS {
-  external String readFileSync(String path, String encoding);
+  external readFileSync(String path, [String encoding]);
 
   external bool existsSync(String path);
 }
@@ -14,6 +16,20 @@ class _FS {
 @JS()
 class _Stderr {
   external void write(String text);
+}
+
+@JS()
+class _SystemError {
+  external String get message;
+  external String get code;
+  external String get syscall;
+  external String get path;
+}
+
+class FileSystemException {
+  final String message;
+
+  FileSystemException._(this.message);
 }
 
 class Stderr {
@@ -35,7 +51,26 @@ external _FS _require(String name);
 
 final _fs = _require("fs");
 
-String readFile(String path) => _fs.readFileSync(path, 'utf8');
+List<int> readFileAsBytes(String path) => _readFile(path) as Uint8List;
+
+String readFileAsString(String path) => _readFile(path, 'utf8') as String;
+
+/// Wraps `fs.readFileSync` to throw a [FileSystemException].
+_readFile(String path, [String encoding]) {
+  try {
+    return _fs.readFileSync(path);
+  } catch (error) {
+    throw new FileSystemException._(_cleanErrorMessage(error as _SystemError));
+  }
+}
+
+/// Cleans up a Node system error's message.
+String _cleanErrorMessage(_SystemError error) {
+  // The error message is of the form "$code: $text, $syscall '$path'". We just
+  // want the text.
+  return error.message.substring("${error.code}: ".length,
+      error.message.length - ", ${error.syscall} '${error.path}'".length);
+}
 
 bool fileExists(String path) => _fs.existsSync(path);