+++ lib/src/io/vm.dart
@@ -4,12 +4,14 @@
 
 import 'dart:io' as io;
 
-export 'dart:io' show exitCode;
+export 'dart:io' show exitCode, FileSystemException;
 
 io.Stdout get stderr => io.stderr;
 
 bool get hasTerminal => io.stdout.hasTerminal;
 
-String readFile(String path) => new io.File(path).readAsStringSync();
+List<int> readFileAsBytes(String path) => new io.File(path).readAsBytesSync();
+
+String readFileAsString(String path) => new io.File(path).readAsStringSync();
 
 bool fileExists(String path) => new io.File(path).existsSync();