+++ lib/src/visitor/perform.dart
@@ -483,10 +483,17 @@ class _PerformVisitor
   }
 
   Value visitForRule(ForRule node) {
-    var from = _addExceptionSpan(node.from.span,
-        () => node.from.accept(this).assertNumber().assertInt());
-    var to = _addExceptionSpan(
-        node.to.span, () => node.to.accept(this).assertNumber().assertInt());
+    var fromNumber = _addExceptionSpan(
+        node.from.span, () => node.from.accept(this).assertNumber());
+    var toNumber = _addExceptionSpan(
+        node.to.span, () => node.to.accept(this).assertNumber());
+
+    var from = _addExceptionSpan(
+        node.from.span,
+        () => fromNumber
+            .coerce(toNumber.numeratorUnits, toNumber.denominatorUnits)
+            .assertInt());
+    var to = _addExceptionSpan(node.to.span, () => toNumber.assertInt());
 
     // TODO: coerce units
     var direction = from > to ? -1 : 1;
@@ -572,7 +579,17 @@ class _PerformVisitor
     }
 
     return _importedFiles.putIfAbsent(path, () {
-      var contents = readFile(path);
+      String contents;
+      try {
+        contents = readSassFile(path);
+      } on SassException catch (error) {
+        var frames = _stack.toList()..add(_stackFrame(import.span));
+        throw new SassRuntimeException(
+            error.message, error.span, new Trace(frames));
+      } on FileSystemException catch (error) {
+        throw _exception(error.message, import.span);
+      }
+
       var url = p.toUri(path);
       return p.extension(path) == '.sass'
           ? new Stylesheet.parseSass(contents, url: url, color: _color)